Drop VS2019 support
