+++ src/ResXManager.VSIX.Compatibility.x64/CustomToolRunner.cs
