+++ src/ResXManager.VSIX.Compatibility.x64/DteConfiguration.cs
