+++ src/ResXManager.VSIX.Compatibility.x64/DteExtensions.cs
