+++ src/ResXManager.VSIX.Compatibility.x64/DteProjectFile.cs
