+++ src/ResXManager.VSIX.Compatibility.x64/DteSolution.cs
