+++ src/ResXManager.VSIX.Compatibility.x64/ErrorListProviderService.cs
