+++ src/ResXManager.VSIX.Compatibility.x64/ICustomToolRunner.cs
