+++ src/ResXManager.VSIX.Compatibility.x64/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+﻿namespace ResXManager.VSIX.Compatibility.x64.Properties;
+
+public static class AssemblyKey
+{
+}
