+++ src/ResXManager.VSIX.Compatibility.x64/Refactorings.cs
