+++ src/ResXManager.VSIX.Compatibility.x64/VsixCompatibility.cs
