--- src/ResXManager.VSIX.Compatibility.x86/FodyWeavers.xml
@@ -1,4 +1,0 @@
-﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-  <AutoProperties />
-  <Throttle />
-</Weavers>
