--- src/ResXManager.VSIX.Compatibility.x86/ResXManager.VSIX.Compatibility.x86.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-    <DisableImplicitNamespaceImports>true</DisableImplicitNamespaceImports>
-  </PropertyGroup>
-  <Import Project="..\ResXManager.VSIX.Compatibility.Shared\ResXManager.VSIX.Compatibility.Shared.projitems" Label="Shared" />
-  <ItemGroup>
-    <PackageReference Include="AutoProperties.Fody" PrivateAssets="all" />
-    <PackageReference Include="Community.VisualStudio.Toolkit" />
-    <PackageReference Include="Microsoft.VisualStudio.SDK" />
-    <PackageReference Include="Fody" PrivateAssets="all" />
-    <PackageReference Include="Throttle.Fody" PrivateAssets="all" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\ResXManager.Infrastructure\ResXManager.Infrastructure.csproj" />
-    <ProjectReference Include="..\ResXManager.Model\ResXManager.Model.csproj" />
-    <ProjectReference Include="..\ResXManager.View\ResXManager.View.csproj" />
-    <ProjectReference Include="..\ResXManager.VSIX.Compatibility\ResXManager.VSIX.Compatibility.csproj" />
-  </ItemGroup>
-</Project>
