+++ src/ResXManager.VSIX/ResXManager.VSIX.csproj
@@ -57,10 +57,9 @@
     <ProjectReference Include="..\ResXManager.VSIX.Compatibility\ResXManager.VSIX.Compatibility.csproj">
       <IncludeOutputGroupsInVSIX>BuiltProjectOutputGroup;BuiltProjectOutputGroupDependencies;GetCopyToOutputDirectoryItems;SatelliteDllsProjectOutputGroup;DebugSymbolsProjectOutputGroup;</IncludeOutputGroupsInVSIX>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\ResXManager.VSIX.Compatibility.x64\ResXManager.VSIX.Compatibility.x64.csproj" ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\ResXManager.VSIX.Compatibility.x86\ResXManager.VSIX.Compatibility.x86.csproj" ReferenceOutputAssembly="false" />
+    <ProjectReference Include="..\ResXManager.VSIX.Compatibility.x64\ResXManager.VSIX.Compatibility.x64.csproj">
+      <IncludeOutputGroupsInVSIX>BuiltProjectOutputGroup;BuiltProjectOutputGroupDependencies;GetCopyToOutputDirectoryItems;SatelliteDllsProjectOutputGroup;DebugSymbolsProjectOutputGroup;</IncludeOutputGroupsInVSIX>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
@@ -72,17 +71,13 @@
     <PackageReference Include="AutoProperties.Fody" PrivateAssets="all" />
     <PackageReference Include="Community.VisualStudio.Toolkit" />
     <PackageReference Include="Fody" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.VisualStudio.SDK" ExcludeAssets="runtime" />
+    <PackageReference Include="Microsoft.VisualStudio.SDK" />
     <PackageReference Include="Microsoft.VSSDK.BuildTools" PrivateAssets="all" />
     <PackageReference Include="PropertyChanged.Fody" PrivateAssets="all" />
     <PackageReference Include="Resourcer.Fody" PrivateAssets="all" />
     <PackageReference Include="Throttle.Fody" PrivateAssets="all" />
     <PackageReference Include="TomsToolbox.Composition.Ninject" />
     <PackageReference Include="VSIX-SdkProjectAdapter" PrivateAssets="all" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="..\ResXManager.VSIX.Compatibility.x86\bin\$(Configuration)\$(TargetFramework)\ResXManager.VSIX.Compatibility.x86.dll" />
-    <EmbeddedResource Include="..\ResXManager.VSIX.Compatibility.x64\bin\$(Configuration)\$(TargetFramework)\ResXManager.VSIX.Compatibility.x64.dll" />
   </ItemGroup>
   <Target Name="PostBuild" AfterTargets="PostBuildEvent" Condition="'$(Configuration)'=='Release'" Outputs="$(SolutionDir)Deploy\ResXManager.VSIX.vsix">
     <MakeDir Directories="$(SolutionDir)Deploy" />
